added unit tests for poller.go, re-added OS conditional to runner.go, removing changes to e2e test
