+++ pkg/diagnose/runner.go
@@ -11,6 +11,7 @@ import (
 	"fmt"
 	"io"
 	"regexp"
+	"runtime"
 	"sort"
 	"strings"
 
@@ -568,5 +569,7 @@ func RegisterConnectivityDatadogEventPlatform(catalog *diagnosis.Catalog) {
 // RegisterPortConflict registers the port-conflict diagnose suite.
 func RegisterPortConflict(catalog *diagnosis.Catalog) {
 	// port-conflict suite available in darwin and linux only for now
-	catalog.Register("port-conflict", func() []diagnosis.Diagnosis { return ports.DiagnosePortSuite() })
+	if runtime.GOOS == "darwin" || runtime.GOOS == "linux" {
+		catalog.Register("port-conflict", func() []diagnosis.Diagnosis { return ports.DiagnosePortSuite() })
+	}
 }
