+++ pkg/util/port/portlist/poller.go
@@ -10,6 +10,7 @@ package portlist
 
 import (
 	"fmt"
+	"sort"
 	"sync"
 
 	"golang.org/x/exp/slices"
@@ -92,4 +93,32 @@ func (p *Poller) getList() (List, error) {
 	var err error
 	p.scratch, err = p.os.AppendListeningPorts(p.scratch[:0])
 	return p.scratch, err
+}
+
+func (a *Port) lessThan(b *Port) bool {
+	if a.Port != b.Port {
+		return a.Port < b.Port
+	}
+	if a.Proto != b.Proto {
+		return a.Proto < b.Proto
+	}
+	return a.Process < b.Process
+}
+
+// sortAndDedup sorts ps in place (by Port.LessThan) and then returns
+// a subset of it with duplicate (Proto, Port) removed.
+func sortAndDedup(ps List) List {
+	sort.Slice(ps, func(i, j int) bool {
+		return (&ps[i]).lessThan(&ps[j])
+	})
+	out := ps[:0]
+	var last Port
+	for _, p := range ps {
+		if last.Proto == p.Proto && last.Port == p.Port {
+			continue
+		}
+		out = append(out, p)
+		last = p
+	}
+	return out
 }
