--- pkg/util/port/portlist/poller_linux_macos.go
@@ -1,38 +1,0 @@
-// Copyright (c) Tailscale Inc & AUTHORS
-// SPDX-License-Identifier: BSD-3-Clause
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2014-present Datadog, Inc.
-
-//go:build darwin || linux
-
-package portlist
-
-import "sort"
-
-func (a *Port) lessThan(b *Port) bool {
-	if a.Port != b.Port {
-		return a.Port < b.Port
-	}
-	if a.Proto != b.Proto {
-		return a.Proto < b.Proto
-	}
-	return a.Process < b.Process
-}
-
-// sortAndDedup sorts ps in place (by Port.LessThan) and then returns
-// a subset of it with duplicate (Proto, Port) removed.
-func sortAndDedup(ps List) List {
-	sort.Slice(ps, func(i, j int) bool {
-		return (&ps[i]).lessThan(&ps[j])
-	})
-	out := ps[:0]
-	var last Port
-	for _, p := range ps {
-		if last.Proto == p.Proto && last.Port == p.Port {
-			continue
-		}
-		out = append(out, p)
-		last = p
-	}
-	return out
-}
