+++ pkg/util/port/portlist/poller_macos_test.go
@@ -85,91 +85,6 @@ func TestPoller(t *testing.T) {
 	}
 }
 
-func TestEqualLessThan(t *testing.T) {
-	tests := []struct {
-		name string
-		a, b Port
-		want bool
-	}{
-		{
-			"Port a < b",
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			Port{Proto: "tcp", Port: 101, Process: "proc1"},
-			true,
-		},
-		{
-			"Port a > b",
-			Port{Proto: "tcp", Port: 101, Process: "proc1"},
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			false,
-		},
-		{
-			"Proto a < b",
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			Port{Proto: "udp", Port: 100, Process: "proc1"},
-			true,
-		},
-		{
-			"Proto a < b",
-			Port{Proto: "udp", Port: 100, Process: "proc1"},
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			false,
-		},
-		{
-			"Process a < b",
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			Port{Proto: "tcp", Port: 100, Process: "proc2"},
-			true,
-		},
-		{
-			"Process a > b",
-			Port{Proto: "tcp", Port: 100, Process: "proc2"},
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			false,
-		},
-		{
-			"Port evaluated first",
-			Port{Proto: "udp", Port: 100, Process: "proc2"},
-			Port{Proto: "tcp", Port: 101, Process: "proc1"},
-			true,
-		},
-		{
-			"Proto evaluated second",
-			Port{Proto: "tcp", Port: 100, Process: "proc2"},
-			Port{Proto: "udp", Port: 100, Process: "proc1"},
-			true,
-		},
-		{
-			"Process evaluated fourth",
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			Port{Proto: "tcp", Port: 100, Process: "proc2"},
-			true,
-		},
-		{
-			"equal",
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			Port{Proto: "tcp", Port: 100, Process: "proc1"},
-			false,
-		},
-	}
-
-	for _, tt := range tests {
-		got := tt.a.lessThan(&tt.b)
-		if got != tt.want {
-			t.Errorf("%s: Equal = %v; want %v", tt.name, got, tt.want)
-		}
-		lessBack := tt.b.lessThan(&tt.a)
-		if got && lessBack {
-			t.Errorf("%s: both a and b report being less than each other", tt.name)
-		}
-		wantEqual := !got && !lessBack
-		gotEqual := tt.a.equal(&tt.b)
-		if gotEqual != wantEqual {
-			t.Errorf("%s: equal = %v; want %v", tt.name, gotEqual, wantEqual)
-		}
-	}
-}
-
 func TestClose(t *testing.T) {
 	var p Poller
 	err := p.Close()
